Removed the need for passing a redirect via cookies.
