+++ api/authentication/src/main/kotlin/uk/co/baconi/oauth/api/authentication/AuthenticationRoute.kt
@@ -1,31 +1,22 @@
 package uk.co.baconi.oauth.api.authentication
 
-import io.ktor.http.*
 import io.ktor.http.ContentType.Application
 import io.ktor.http.HttpStatusCode.Companion.BadRequest
-import io.ktor.http.HttpStatusCode.Companion.NotAcceptable
 import io.ktor.http.HttpStatusCode.Companion.OK
 import io.ktor.http.HttpStatusCode.Companion.Unauthorized
 import io.ktor.http.HttpStatusCode.Companion.UnsupportedMediaType
 import io.ktor.server.application.*
-import io.ktor.server.html.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
-import io.ktor.server.util.*
 import uk.co.baconi.oauth.api.authentication.AuthenticationRequest.InvalidField
 import uk.co.baconi.oauth.api.authentication.AuthenticationRequest.Valid
 import uk.co.baconi.oauth.api.common.authentication.CustomerAuthentication
 import uk.co.baconi.oauth.api.common.authentication.CustomerAuthentication.Failure
 import uk.co.baconi.oauth.api.common.authentication.CustomerAuthentication.Success
 import uk.co.baconi.oauth.api.common.authentication.CustomerAuthenticationService
-import uk.co.baconi.oauth.api.common.html.ReactTemplate.reactPage
-import uk.co.baconi.oauth.api.common.location.Location
-import java.util.*
 import kotlin.time.Duration.Companion.hours
-import kotlin.time.Duration.Companion.minutes
 
 const val COOKIE_CSRF = "Authenticate-CSRF"
-const val COOKIE_REDIRECT_URI = "Authenticate-Redirect-Uri"
 const val COOKIE_CUSTOMER = "Authenticated-Customer"
 
 interface AuthenticationRoute : AuthenticationRequestValidation {
@@ -36,19 +27,7 @@ interface AuthenticationRoute : AuthenticationRequestValidation {
 
         application.log.info("Registering the AuthenticationRoute.authentication() routes")
 
-        val bundleLocation = url {
-            takeFrom(Location.Assets.baseUrl)
-            path("/assets/js/authentication.js")
-        }
-
-        application.log.debug("Authentication asset location: {}", bundleLocation)
-
         route("/authentication") {
-            accept(Application.Json) {
-                get {
-                    call.respond(NotAcceptable)
-                }
-            }
             contentType(Application.Json) {
                 post {
                     when (val request = call.validateAuthenticationRequest()) {
@@ -76,28 +55,6 @@ interface AuthenticationRoute : AuthenticationRequestValidation {
                             }
                         }
                     }
-                }
-            }
-            get {
-                // TODO - Handle already being logged in, do we redirect, pop out a new UI or assume you're logging in as someone new?
-                call.respondHtml(OK) {
-
-                    val csrfToken = UUID.randomUUID()
-
-                    // TODO - Convert to signed client session or switch to passed JWT.
-                    call.response.cookies.append(
-                        name = COOKIE_CSRF,
-                        value = csrfToken.toString(),
-                        maxAge = 30.minutes.inWholeSeconds,
-                        //secure = true, // TODO - Enable if behind TLS, may need https://ktor.io/docs/forward-headers.html
-                        httpOnly = true,
-                    )
-
-                    val redirectUri = call.request.cookies[COOKIE_REDIRECT_URI]
-
-                    application.log.debug("CSRF: {}, Redirect URI: {}", csrfToken, redirectUri)
-
-                    reactPage(title = "Login Page", reactSource = bundleLocation, csrfToken, redirectUri)
                 }
             }
             post {
